requirements.txt
python-dotenv
requests
python-pptx
sendgrid
